backend/requirements.txt
fastapi==0.115.0
pydantic-settings==2.5.2
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
greenlet==3.1.1
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
redis==5.1.1
python-dotenv==1.0.1
pillow==10.4.0
aiofiles==24.1.0
websockets==13.1
email-validator==2.2.0
httpx==0.27.2
livekit-api==0.7.2
